refactor: use is_private and workspace membership to determine eligibility to reopen
