+++ lib/modules/platform/bitbucket/types.ts
@@ -16,6 +16,7 @@ export interface Config {
   prList: Pr[];
   repository: string;
   ignorePrAuthor: boolean;
+  isPrivate: boolean;
 }
 
 export interface PagedResult<T = any> {
@@ -32,6 +33,7 @@ export interface RepoInfo {
   mergeMethod: string;
   has_issues: boolean;
   uuid: string;
+  isPrivate: boolean;
 }
 
 export interface RepoBranchingModel {
@@ -63,6 +65,7 @@ export interface RepoInfoBody {
   has_issues: boolean;
   uuid: string;
   full_name: string;
+  is_private: boolean;
 }
 
 export interface PrResponse {
